move check_auto_ext to formats.py
